pyproject.toml
[tool.poetry]
name = "irrd"
version = "0.0.1"
description = "IRRd Test"
authors = ["MattKobayashi <matthew@kobayashi.au>"]
package-mode = false

[tool.poetry.dependencies]
python = "3.13.3"
psycopg2-binary = "2.9.10"
PyYAML = "6.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core==2.1.2"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
